refactor(train): extract profiler helpers from util.py into profiler.py
